requirements.txt
python-telegram-bot==20.7
selenium==4.15.2
pandas==2.1.3
openpyxl==3.1.2